pyproject.toml
[tool.poetry]
name = "nonebot-plugin-splatoon3-schedule"
version = "1.9.4"
description = "一个基于nonebot2框架的splatoon3游戏日程查询插件"
authors = ["cypas <ayano05@outlook.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_splatoon3_schedule"}]

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4="^4.12.2"
cfscrape="^2.1.1"
httpx="^0.24.1"
nonebot2="^2.2.0"
pydantic = ">=1.0.0,<3.0.0,!=2.5.0,!=2.5.1"
Pillow="9.5.0"
urllib3="1.26.16"
playwright = "^1.40.0"
nonebot-adapter-onebot = "^2.4.1"
nonebot-adapter-telegram= "^0.1.0b16"
nonebot-adapter-kaiheila = "^0.3.1"
nonebot-adapter-qq = "^1.4.1"
nonebot-plugin-apscheduler = "^0.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
